In compact trie, add missing %LOOKUP method
The case that was not covered is the following: The traversal reaches
an interior node and a proper prefix key of the query string matches the key
one child cell. Since the prefix is proper, there is some of the query
string "left". Now, if value of the child cell is an omitted leaf
node, that is one or more compact entries, %LOOKUP is called like

  (%lookup FUNCTION STRING NON-ZERO-SUFFIX INTEGER-OR-CONS)

The correct behavior is to stop the traversal since the compact
entries would match only if the suffix was zero.

This change adds a method that is applicable in this case and stops
the traversal.